\name{mdata}
\alias{mdata}
\title{
Dataset with missing values to illustrate use of package norm
}
\description{
Household survey with missing values. See Schafer~(1997).
}
\references{
Schafer, J.L.(1997) \emph{Analysis of Incomplete Multivariate Data},
Chapman & Hall, London. ISBN: 0412040611
}
\keyword{
datasets
}
