% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{A Normalize Function}
\usage{
normalize(data)
}
\arguments{
\item{Data}{= Dataframe you need to normalize.}
}
\description{
This function allows you to normalize a dataframe.
}
\examples{
normalize()
}
\keyword{Normalize}

